+++ src/main/java/net/succ/solar_punk/block/custom/TurbineCasingGlassBlock.java
@@ -1,0 +1,10 @@
+package net.succ.solar_punk.block.custom;
+
+import com.simibubi.create.content.equipment.wrench.IWrenchable;
+import net.minecraft.world.level.block.Block;
+
+public class TurbineCasingGlassBlock extends Block implements IWrenchable {
+    public TurbineCasingGlassBlock(Properties properties) {
+        super(properties);
+    }
+}